# internal/generated/api.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.0 DO NOT EDIT.
package generated

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerJWTScopes    = "BearerJWT.Scopes"
	BearerPASETOScopes = "BearerPASETO.Scopes"
)

// Defines values for CreateMovementRequestType.
const (
	CreateMovementRequestTypeCredit CreateMovementRequestType = "credit"
	CreateMovementRequestTypeDebit  CreateMovementRequestType = "debit"
)

// Defines values for MovementType.
const (
	MovementTypeCredit MovementType = "credit"
	MovementTypeDebit  MovementType = "debit"
)

// Defines values for TransferStatus.
const (
	Completed TransferStatus = "completed"
	Failed    TransferStatus = "failed"
	Pending   TransferStatus = "pending"
)

// APIError defines model for APIError.
type APIError struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// AuthResponse defines model for AuthResponse.
type AuthResponse struct {
	// ExpiresIn Token TTL in seconds (handler currently hard-codes 3600)
	ExpiresIn int32 `json:"expires_in"`

	// Token JWT access token
	Token string `json:"token"`
}

// BalanceResponse defines model for BalanceResponse.
type BalanceResponse struct {
	AccountId UUID `json:"account_id"`

	// Balance Balance formatted as string (decimal.Decimal.String()).
	Balance  string `json:"balance"`
	Currency string `json:"currency"`
}

// CreateMovementRequest defines model for CreateMovementRequest.
type CreateMovementRequest struct {
	// Amount Decimal encoded as string (shopspring/decimal)
	Amount      DecimalString             `json:"amount"`
	Description *string                   `json:"description,omitempty"`
	Type        CreateMovementRequestType `json:"type"`
}

// CreateMovementRequestType defines model for CreateMovementRequest.Type.
type CreateMovementRequestType string

// DateTime defines model for DateTime.
type DateTime = time.Time

// DecimalString Decimal encoded as string (shopspring/decimal)
type DecimalString = string

// ErrorResponse Current error envelope from `internal/util/errors.go`.
// Note: for non-*util.APIError errors, the server responds with code=500 and message="internal server error".
type ErrorResponse struct {
	Error APIError `json:"error"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
}

// Movement Mirrors `internal/model.Movement` JSON.
// NOTE: in Go it serializes `amount` as decimal (shopspring/decimal) which is typically a JSON string/number depending on config.
// TODO: confirm runtime JSON encoding for decimal.Decimal and adjust if needed.
type Movement struct {
	AccountId UUID `json:"account_id"`

	// Amount Decimal encoded as string (shopspring/decimal)
	Amount      DecimalString `json:"amount"`
	Description string        `json:"description"`
	Id          int64         `json:"id"`
	OccurredAt  DateTime      `json:"occurred_at"`
	Type        MovementType  `json:"type"`
}

// MovementType defines model for Movement.Type.
type MovementType string

// PaginatedMovementsResponse Concrete shape of `util.PaginatedResponse` as returned by `MovementService.GetByAccountID()`.
type PaginatedMovementsResponse struct {
	Data       []Movement     `json:"data"`
	Pagination PaginationMeta `json:"pagination"`
}

// PaginatedTransfersResponse Concrete shape of `util.PaginatedResponse` as returned by `TransferService.GetByAccountID()`.
type PaginatedTransfersResponse struct {
	Data       []Transfer     `json:"data"`
	Pagination PaginationMeta `json:"pagination"`
}

// PaginationMeta defines model for PaginationMeta.
type PaginationMeta struct {
	CurrentPage int32 `json:"current_page"`
	PerPage     int32 `json:"per_page"`
	TotalItems  int32 `json:"total_items"`
	TotalPages  int32 `json:"total_pages"`
}

// SignUpRequest defines model for SignUpRequest.
type SignUpRequest struct {
	Email      openapi_types.Email `json:"email"`
	FirstName  string              `json:"first_name"`
	FiscalCode string              `json:"fiscal_code"`
	LastName   string              `json:"last_name"`
	Password   string              `json:"password"`
	Username   string              `json:"username"`
}

// Transfer defines model for Transfer.
type Transfer struct {
	// Amount Decimal encoded as string (shopspring/decimal)
	Amount      DecimalString  `json:"amount"`
	CompletedAt *time.Time     `json:"completed_at"`
	FromAccount UUID           `json:"from_account"`
	Id          int64          `json:"id"`
	InitiatedAt DateTime       `json:"initiated_at"`
	Status      TransferStatus `json:"status"`
	ToAccount   UUID           `json:"to_account"`
}

// TransferStatus defines model for Transfer.Status.
type TransferStatus string

// TransferRequest defines model for TransferRequest.
type TransferRequest struct {
	// Amount Decimal encoded as string (shopspring/decimal)
	Amount      DecimalString `json:"amount"`
	Description *string       `json:"description,omitempty"`
	ToAccount   UUID          `json:"to_account"`
}

// UUID defines model for UUID.
type UUID = openapi_types.UUID

// User defines model for User.
type User struct {
	CreatedAt  DateTime            `json:"created_at"`
	Email      openapi_types.Email `json:"email"`
	FirstName  string              `json:"first_name"`
	FiscalCode string              `json:"fiscal_code"`
	Id         UUID                `json:"id"`
	LastName   string              `json:"last_name"`
	UpdatedAt  DateTime            `json:"updated_at"`
	Username   string              `json:"username"`
}

// LimitParam defines model for LimitParam.
type LimitParam = int

// OAuthCodeParam defines model for OAuthCodeParam.
type OAuthCodeParam = string

// OAuthStateParam defines model for OAuthStateParam.
type OAuthStateParam = string

// PageParam defines model for PageParam.
type PageParam = int

// BadRequestError Current error envelope from `internal/util/errors.go`.
// Note: for non-*util.APIError errors, the server responds with code=500 and message="internal server error".
type BadRequestError = ErrorResponse

// InternalServerError Current error envelope from `internal/util/errors.go`.
// Note: for non-*util.APIError errors, the server responds with code=500 and message="internal server error".
type InternalServerError = ErrorResponse

// UnauthorizedError Current error envelope from `internal/util/errors.go`.
// Note: for non-*util.APIError errors, the server responds with code=500 and message="internal server error".
type UnauthorizedError = ErrorResponse

// AccountsListMovementsParams defines parameters for AccountsListMovements.
type AccountsListMovementsParams struct {
	// Page Page number (default: 1)
	Page *PageParam `form:"page,omitempty" json:"page,omitempty"`

	// Limit Items per page (default: 10, max: 100)
	Limit *LimitParam `form:"limit,omitempty" json:"limit,omitempty"`
}

// AuthGoogleCallbackParams defines parameters for AuthGoogleCallback.
type AuthGoogleCallbackParams struct {
	// Code OAuth code
	Code OAuthCodeParam `form:"code" json:"code"`

	// State CSRF state
	State OAuthStateParam `form:"state" json:"state"`
}

// TransfersListParams defines parameters for TransfersList.
type TransfersListParams struct {
	// Page Page number (default: 1)
	Page *PageParam `form:"page,omitempty" json:"page,omitempty"`

	// Limit Items per page (default: 10, max: 100)
	Limit *LimitParam `form:"limit,omitempty" json:"limit,omitempty"`
}

// AccountsCreateMovementJSONRequestBody defines body for AccountsCreateMovement for application/json ContentType.
type AccountsCreateMovementJSONRequestBody = CreateMovementRequest

// AuthLoginJSONRequestBody defines body for AuthLogin for application/json ContentType.
type AuthLoginJSONRequestBody = LoginRequest

// AuthSignUpJSONRequestBody defines body for AuthSignUp for application/json ContentType.
type AuthSignUpJSONRequestBody = SignUpRequest

// TransfersCreateJSONRequestBody defines body for TransfersCreate for application/json ContentType.
type TransfersCreateJSONRequestBody = TransferRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get account balance
	// (GET /api/v1/accounts/balance)
	AccountsGetBalance(c *gin.Context)
	// List account movements (paginated)
	// (GET /api/v1/accounts/movements)
	AccountsListMovements(c *gin.Context, params AccountsListMovementsParams)
	// Create account movement
	// (POST /api/v1/accounts/movements)
	AccountsCreateMovement(c *gin.Context)
	// Start Google OAuth flow
	// (GET /api/v1/auth/google)
	AuthGoogle(c *gin.Context)
	// Handle Google OAuth callback
	// (GET /api/v1/auth/google/callback)
	AuthGoogleCallback(c *gin.Context, params AuthGoogleCallbackParams)
	// Login
	// (POST /api/v1/auth/login)
	AuthLogin(c *gin.Context)
	// Register a new user
	// (POST /api/v1/auth/signup)
	AuthSignUp(c *gin.Context)
	// List transfers (paginated)
	// (GET /api/v1/transfers)
	TransfersList(c *gin.Context, params TransfersListParams)
	// Create a transfer
	// (POST /api/v1/transfers)
	TransfersCreate(c *gin.Context)
	// Health check
	// (GET /health)
	HealthCheck(c *gin.Context)
	// Prometheus metrics
	// (GET /metrics)
	Metrics(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// AccountsGetBalance operation middleware
func (siw *ServerInterfaceWrapper) AccountsGetBalance(c *gin.Context) {

	c.Set(BearerJWTScopes, []string{})

	c.Set(BearerPASETOScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AccountsGetBalance(c)
}

// AccountsListMovements operation middleware
func (siw *ServerInterfaceWrapper) AccountsListMovements(c *gin.Context) {

	var err error

	c.Set(BearerJWTScopes, []string{})

	c.Set(BearerPASETOScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params AccountsListMovementsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", c.Request.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", c.Request.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AccountsListMovements(c, params)
}

// AccountsCreateMovement operation middleware
func (siw *ServerInterfaceWrapper) AccountsCreateMovement(c *gin.Context) {

	c.Set(BearerJWTScopes, []string{})

	c.Set(BearerPASETOScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AccountsCreateMovement(c)
}

// AuthGoogle operation middleware
func (siw *ServerInterfaceWrapper) AuthGoogle(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AuthGoogle(c)
}

// AuthGoogleCallback operation middleware
func (siw *ServerInterfaceWrapper) AuthGoogleCallback(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params AuthGoogleCallbackParams

	// ------------- Required query parameter "code" -------------

	if paramValue := c.Query("code"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument code is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "code", c.Request.URL.Query(), &params.Code)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter code: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "state" -------------

	if paramValue := c.Query("state"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument state is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "state", c.Request.URL.Query(), &params.State)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter state: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AuthGoogleCallback(c, params)
}

// AuthLogin operation middleware
func (siw *ServerInterfaceWrapper) AuthLogin(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AuthLogin(c)
}

// AuthSignUp operation middleware
func (siw *ServerInterfaceWrapper) AuthSignUp(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AuthSignUp(c)
}

// TransfersList operation middleware
func (siw *ServerInterfaceWrapper) TransfersList(c *gin.Context) {

	var err error

	c.Set(BearerJWTScopes, []string{})

	c.Set(BearerPASETOScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params TransfersListParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", c.Request.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", c.Request.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.TransfersList(c, params)
}

// TransfersCreate operation middleware
func (siw *ServerInterfaceWrapper) TransfersCreate(c *gin.Context) {

	c.Set(BearerJWTScopes, []string{})

	c.Set(BearerPASETOScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.TransfersCreate(c)
}

// HealthCheck operation middleware
func (siw *ServerInterfaceWrapper) HealthCheck(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.HealthCheck(c)
}

// Metrics operation middleware
func (siw *ServerInterfaceWrapper) Metrics(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Metrics(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/api/v1/accounts/balance", wrapper.AccountsGetBalance)
	router.GET(options.BaseURL+"/api/v1/accounts/movements", wrapper.AccountsListMovements)
	router.POST(options.BaseURL+"/api/v1/accounts/movements", wrapper.AccountsCreateMovement)
	router.GET(options.BaseURL+"/api/v1/auth/google", wrapper.AuthGoogle)
	router.GET(options.BaseURL+"/api/v1/auth/google/callback", wrapper.AuthGoogleCallback)
	router.POST(options.BaseURL+"/api/v1/auth/login", wrapper.AuthLogin)
	router.POST(options.BaseURL+"/api/v1/auth/signup", wrapper.AuthSignUp)
	router.GET(options.BaseURL+"/api/v1/transfers", wrapper.TransfersList)
	router.POST(options.BaseURL+"/api/v1/transfers", wrapper.TransfersCreate)
	router.GET(options.BaseURL+"/health", wrapper.HealthCheck)
	router.GET(options.BaseURL+"/metrics", wrapper.Metrics)
}